+++ packages/docs/src/functions/getDataOptionsFromQueryString.ts
@@ -1,7 +1,9 @@
 import { IndexedObject } from '@cnamts/vue-dot/src/types';
-
-import { EventQueryStringParameters } from '@netlify/functions/src/function/event';
 import { DataOptions } from 'vuetify';
+
+interface EventQueryStringParameters {
+	[name: string]: string | undefined;
+}
 
 type ParsedValue = string | boolean | number;
 type Params = ParsedValue[] | string | boolean | number;
@@ -14,8 +16,12 @@ function parseValue(value: string): ParsedValue {
 	}
 }
 
-export function getDataOptionsFromQueryString(queryString: EventQueryStringParameters): DataOptions {
+export function getDataOptionsFromQueryString(queryString: EventQueryStringParameters | null): DataOptions {
 	const params: IndexedObject<Params> = {};
+
+	if (queryString === null) {
+		return params as unknown as DataOptions;
+	}
 
 	Object.entries(queryString).map(([key, value]) => {
 		if (key.match(/\[\]/)) {